+++ core/model/descriptors/backup-s3.mjson
@@ -1,12 +1,4 @@
 {
-  "blueprint_backups3_access_key": {
-    "prototype": "montage/core/meta/property-blueprint",
-    "properties": {
-      "name": "access_key",
-      "blueprint": {"@": "root"},
-      "valueType": "String"
-    }
-  },
   "blueprint_backups3_bucket": {
     "prototype": "montage/core/meta/property-blueprint",
     "properties": {
@@ -23,18 +15,10 @@
       "valueType": "String"
     }
   },
-  "blueprint_backups3_region": {
+  "blueprint_backups3_peer": {
     "prototype": "montage/core/meta/property-blueprint",
     "properties": {
-      "name": "region",
-      "blueprint": {"@": "root"},
-      "valueType": "String"
-    }
-  },
-  "blueprint_backups3_secret_key": {
-    "prototype": "montage/core/meta/property-blueprint",
-    "properties": {
-      "name": "secret_key",
+      "name": "peer",
       "blueprint": {"@": "root"},
       "valueType": "String"
     }
@@ -54,11 +38,9 @@
     "properties": {
       "name": "BackupS3",
       "propertyBlueprints": [
-        {"@": "blueprint_backups3_access_key"},
         {"@": "blueprint_backups3_bucket"},
         {"@": "blueprint_backups3_folder"},
-        {"@": "blueprint_backups3_region"},
-        {"@": "blueprint_backups3_secret_key"},
+        {"@": "blueprint_backups3_peer"},
         {"@": "blueprint_backups3_type"}
       ]
     }
