+++ core/model/descriptors/freenas-credentials.mjson
@@ -1,4 +1,12 @@
 {
+  "blueprint_freenascredentials_address": {
+    "prototype": "montage/core/meta/property-blueprint",
+    "properties": {
+      "name": "address",
+      "blueprint": {"@": "root"},
+      "valueType": "String"
+    }
+  },
   "blueprint_freenascredentials_hostkey": {
     "prototype": "montage/core/meta/property-blueprint",
     "properties": {
@@ -38,6 +46,7 @@
     "properties": {
       "name": "FreenasCredentials",
       "propertyBlueprints": [
+        {"@": "blueprint_freenascredentials_address"},
         {"@": "blueprint_freenascredentials_hostkey"},
         {"@": "blueprint_freenascredentials_port"},
         {"@": "blueprint_freenascredentials_pubkey"},
