--- core/model/descriptors/service-ipfs.mjson
@@ -1,48 +1,0 @@
-{
-  "blueprint_serviceipfs_enable": {
-    "prototype": "montage/core/meta/property-blueprint",
-    "properties": {
-      "name": "enable",
-      "blueprint": {"@": "root"},
-      "valueType": "boolean"
-    }
-  },
-  "blueprint_serviceipfs_path": {
-    "prototype": "montage/core/meta/property-blueprint",
-    "properties": {
-      "name": "path",
-      "blueprint": {"@": "root"},
-      "valueType": "String"
-    }
-  },
-  "blueprint_serviceipfs_type": {
-    "prototype": "montage/core/meta/property-blueprint",
-    "properties": {
-      "name": "type",
-      "blueprint": {"@": "root"},
-      "enumValues": [
-        "service-ipfs"
-      ]
-    }
-  },
-  "blueprint_serviceipfs_webui": {
-    "prototype": "montage/core/meta/property-blueprint",
-    "properties": {
-      "name": "webui",
-      "blueprint": {"@": "root"},
-      "valueType": "boolean"
-    }
-  },
-  "root": {
-    "prototype": "core/model/model-descriptor",
-    "properties": {
-      "name": "ServiceIpfs",
-      "propertyBlueprints": [
-        {"@": "blueprint_serviceipfs_enable"},
-        {"@": "blueprint_serviceipfs_path"},
-        {"@": "blueprint_serviceipfs_type"},
-        {"@": "blueprint_serviceipfs_webui"}
-      ]
-    }
-  }
-}
