+++ core/model/descriptors/ssh-credentials.mjson
@@ -1,4 +1,12 @@
 {
+  "blueprint_sshcredentials_address": {
+    "prototype": "montage/core/meta/property-blueprint",
+    "properties": {
+      "name": "address",
+      "blueprint": {"@": "root"},
+      "valueType": "String"
+    }
+  },
   "blueprint_sshcredentials_hostkey": {
     "prototype": "montage/core/meta/property-blueprint",
     "properties": {
@@ -23,10 +31,10 @@
       "valueType": "number"
     }
   },
-  "blueprint_sshcredentials_pubkey": {
+  "blueprint_sshcredentials_privkey": {
     "prototype": "montage/core/meta/property-blueprint",
     "properties": {
-      "name": "pubkey",
+      "name": "privkey",
       "blueprint": {"@": "root"},
       "valueType": "String"
     }
@@ -54,10 +62,11 @@
     "properties": {
       "name": "SshCredentials",
       "propertyBlueprints": [
+        {"@": "blueprint_sshcredentials_address"},
         {"@": "blueprint_sshcredentials_hostkey"},
         {"@": "blueprint_sshcredentials_password"},
         {"@": "blueprint_sshcredentials_port"},
-        {"@": "blueprint_sshcredentials_pubkey"},
+        {"@": "blueprint_sshcredentials_privkey"},
         {"@": "blueprint_sshcredentials_type"},
         {"@": "blueprint_sshcredentials_username"}
       ]
