+++ core/model/models/backup-ssh.js
@@ -14,56 +14,17 @@ exports.BackupSsh = AbstractModel.specialize({
             return this._directory;
         }
     },
-    _hostkey: {
+    _peer: {
         value: null
     },
-    hostkey: {
+    peer: {
         set: function (value) {
-            if (this._hostkey !== value) {
-                this._hostkey = value;
+            if (this._peer !== value) {
+                this._peer = value;
             }
         },
         get: function () {
-            return this._hostkey;
-        }
-    },
-    _hostport: {
-        value: null
-    },
-    hostport: {
-        set: function (value) {
-            if (this._hostport !== value) {
-                this._hostport = value;
-            }
-        },
-        get: function () {
-            return this._hostport;
-        }
-    },
-    _password: {
-        value: null
-    },
-    password: {
-        set: function (value) {
-            if (this._password !== value) {
-                this._password = value;
-            }
-        },
-        get: function () {
-            return this._password;
-        }
-    },
-    _privkey: {
-        value: null
-    },
-    privkey: {
-        set: function (value) {
-            if (this._privkey !== value) {
-                this._privkey = value;
-            }
-        },
-        get: function () {
-            return this._privkey;
+            return this._peer;
         }
     },
     _type: {
@@ -78,19 +39,6 @@ exports.BackupSsh = AbstractModel.specialize({
         get: function () {
             return this._type;
         }
-    },
-    _username: {
-        value: null
-    },
-    username: {
-        set: function (value) {
-            if (this._username !== value) {
-                this._username = value;
-            }
-        },
-        get: function () {
-            return this._username;
-        }
     }
 }, {
     propertyBlueprints: {
@@ -100,27 +48,11 @@ exports.BackupSsh = AbstractModel.specialize({
             valueType: "String"
         }, {
             mandatory: false,
-            name: "hostkey",
-            valueType: "String"
-        }, {
-            mandatory: false,
-            name: "hostport",
-            valueType: "String"
-        }, {
-            mandatory: false,
-            name: "password",
-            valueType: "String"
-        }, {
-            mandatory: false,
-            name: "privkey",
+            name: "peer",
             valueType: "String"
         }, {
             mandatory: false,
             name: "type"
-        }, {
-            mandatory: false,
-            name: "username",
-            valueType: "String"
         }]
     }
 });
