+++ core/model/models/freenas-credentials.js
@@ -1,6 +1,19 @@
 var AbstractModel = require("core/model/abstract-model").AbstractModel;
 
 exports.FreenasCredentials = AbstractModel.specialize({
+    _address: {
+        value: null
+    },
+    address: {
+        set: function (value) {
+            if (this._address !== value) {
+                this._address = value;
+            }
+        },
+        get: function () {
+            return this._address;
+        }
+    },
     _hostkey: {
         value: null
     },
@@ -56,6 +69,10 @@ exports.FreenasCredentials = AbstractModel.specialize({
 }, {
     propertyBlueprints: {
         value: [{
+            mandatory: false,
+            name: "address",
+            valueType: "String"
+        }, {
             mandatory: false,
             name: "hostkey",
             valueType: "String"
